Add EventFilterStructs.h, space some things out
