+++ JG/nvse_plugin_example.vcxproj.filters
@@ -218,6 +218,9 @@
     <ClInclude Include="events\LambdaVariableContext.h">
       <Filter>events</Filter>
     </ClInclude>
+    <ClInclude Include="events\EventFilterStructs.h">
+      <Filter>events</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <None Include="exports.def" />
